refactor: Type Definitions and imports
