+++ definitions/clients.d.ts
@@ -1,0 +1,17 @@
+interface Client {
+  socketId: string;
+  name: string;
+  role: string;
+}
+
+interface RenderClient {
+  client: Client;
+  prefix?: string;
+  sufix?: string;
+  badges: Badge[]
+}
+
+interface Badge {
+  id: string;
+  onClick: () => void;
+}
