+++ definitions/intervals.d.ts
@@ -1,0 +1,7 @@
+interface PluginInjectIntervals {
+  syncButton: NodeJS.Timeout | null;
+  leaveButton: NodeJS.Timeout | null
+  removeUpnext: NodeJS.Timeout | null
+  queueInject: NodeJS.Timeout | null
+  queueAddButton: NodeJS.Timeout | null
+}
