+++ definitions/options.d.ts
@@ -1,0 +1,9 @@
+interface PlayerOptions {
+  connection: ServerConnectionOptions;
+}
+
+interface ServerConnectionOptions {
+  protocol: string;
+  host: string;
+  port: string;
+}
