+++ definitions/plugin.d.ts
@@ -1,0 +1,10 @@
+interface Reaction {
+  id: string;
+  symbol: string;
+  tooltip: string;
+  text: string;
+}
+
+interface Settings {
+  showReactions: boolean;
+}
