+++ definitions/yt-app.d.ts
@@ -1,0 +1,16 @@
+interface YTApp {
+  onYtNavigate_: (e: Endpoint) => void;
+}
+
+interface Endpoint {
+  detail: {
+    endpoint: {
+      watchEndpoint: {
+        videoId: any;
+      }
+    },
+    params: {
+      syncId: string;
+    }
+  }
+}
