+++ src/util/yt-html.ts
@@ -1,6 +1,21 @@
 import { InjectAction } from '../enum/action';
-import Client from '../model/client';
-import { QUEUE_CONTAINER_SELECTOR, ROOM_INFO_CONTAINER_SELECTOR, REACTIONS_CONTAINER_SELECTOR, REACTION_TIME_TILL_REMOVE, REACTION_FADE_IN_TIME, getReactionId, getReactionTooltipId, REACTION_OVERLAY_ID, ROOM_INFO_CONTAINER_ID, AUTOPLAY_TOGGLE_ID, AUTOPLAY_TOGGLE_TOOLTIP_ID, REACTION_TOGGLE_ID, REACTION_TOGGLE_TOOLTIP_ID, REACTION_CONTAINER_ID, QUEUE_CONTAINER_ID } from './consts';
+import {
+  QUEUE_CONTAINER_SELECTOR,
+  ROOM_INFO_CONTAINER_SELECTOR,
+  REACTIONS_CONTAINER_SELECTOR,
+  REACTION_TIME_TILL_REMOVE,
+  REACTION_FADE_IN_TIME,
+  getReactionId,
+  getReactionTooltipId,
+  REACTION_OVERLAY_ID,
+  ROOM_INFO_CONTAINER_ID,
+  AUTOPLAY_TOGGLE_ID,
+  AUTOPLAY_TOGGLE_TOOLTIP_ID,
+  REACTION_TOGGLE_ID,
+  REACTION_TOGGLE_TOOLTIP_ID,
+  REACTION_CONTAINER_ID,
+  QUEUE_CONTAINER_ID
+} from './consts';
 
 export default class YTHTMLUtil {
   /**
